hard drive tooltips (if initialized; added mkdir to fs driver on lua side; fixed recursive delete (folders)

+++ li/cil/oc/api/detail/FileSystemAPI.scala
@@ -7,7 +7,7 @@ import li.cil.oc.api.network.Node
 trait FileSystemAPI {
   def fromClass(clazz: Class[_], domain: String, root: String): Option[FileSystem]
 
-  def fromSaveDir(root: String, capacity: Int): Option[FileSystem]
+  def fromSaveDir(root: String, capacity: Long): Option[FileSystem]
 
   def asNode(fs: FileSystem): Option[Node]
 }
